refactor: view.js의 비즈니스 로직 utils.js로 추출
